Start some clean up
